rearrange the game logic
